refactor: move cairo into the rpc crate
