+++ crates/pathfinder/Cargo.toml
@@ -52,7 +52,7 @@ sha3 = "0.9"
 stark_hash = { path = "../stark_hash" }
 starknet-gateway-client = { path = "../gateway-client" }
 starknet-gateway-types = { path = "../gateway-types" }
-tempfile = "3"
+# tempfile = "3"
 thiserror = "1.0.30"
 tokio = { version = "1.11.0", features = ["process"] }
 tokio-retry = "0.3.0"
