+++ crates/pathfinder/src/lib.rs
@@ -1,5 +1,4 @@
 #![deny(rust_2018_idioms)]
 
-pub mod cairo;
 pub mod monitoring;
 pub mod state;
