+++ crates/rpc/Cargo.toml
@@ -8,10 +8,14 @@ rust-version = "1.62"
 
 [dependencies]
 anyhow = "1.0.44"
+base64 = "0.13.0"
+flate2 = "1.0.23"
+futures = { version = "0.3", default-features = false, features = ["std"] }
 jsonrpsee = { git = "https://github.com/eqlabs/jsonrpsee", branch = "start_with_paths", default-features = false, features = [
     "server",
 ] }
 metrics = "0.20.1"
+# pathfinder-cairo = { path = "../cairo" }
 pathfinder-common = { path = "../common" }
 pathfinder-ethereum = { path = "../ethereum" }
 pathfinder-serde = { path = "../serde" }
@@ -30,17 +34,16 @@ web3 = "0.18.0"
 
 [dev-dependencies]
 assert_matches = "1.5.0"
-base64 = "0.13.0"
 bytes = "1.1.0"
-flate2 = "1.0.23"
 hex = "0.4.3"
+hex-literal = "0.3"
 jsonrpsee = { git = "https://github.com/eqlabs/jsonrpsee", branch = "start_with_paths", default-features = false, features = [
     "server",
     "async-client",
 ] }
-futures = { version = "0.3", default-features = false, features = ["std"] }
 lazy_static = "1.4.0"
 pretty_assertions = "1.0.0"
+reqwest = { version = "0.11.4", features = ["json"] }
 rusqlite = { version = "0.27.0", features = ["bundled"] }
 serde_json = { version = "1.0.68", features = ["arbitrary_precision", "raw_value"] }
 stark_hash = { path = "../stark_hash" }
